+++ spec/acceptance/class_spec.rb
@@ -10,7 +10,7 @@ describe 'puppet_agent class' do
 
     it 'should work idempotently with no errors' do
       pp = <<-EOS
-      class { 'puppet_agent': package_version => '5.5.16', collection => 'puppet5' }
+      class { 'puppet_agent': package_version => '5.5.21', collection => 'puppet5' }
       EOS
 
       apply_manifest(pp, :catch_failures => true)
@@ -140,7 +140,7 @@ describe 'puppet_agent class' do
     context 'with mcollective configured' do
       before(:all) {
         setup_puppet_on default, :mcollective => true, :agent => true
-        manifest = 'class { "puppet_agent": package_version => "5.5.16", collection => "puppet5", service_names => ["mcollective"] }'
+        manifest = 'class { "puppet_agent": package_version => "5.5.21", collection => "puppet5", service_names => ["mcollective"] }'
         pp = "file { '#{master.puppet['codedir']}/environments/production/manifests/site.pp': ensure => file, content => '#{manifest}' }"
         apply_manifest_on(master, pp, :catch_failures => true)
       }
@@ -197,7 +197,7 @@ describe 'puppet_agent class' do
         end
 
         let(:manifest) do <<-EOS
-      class { 'puppet_agent': package_version => '5.5.16', collection => 'puppet5', before => File['/tmp/a'] }
+      class { 'puppet_agent': package_version => '5.5.21', collection => 'puppet5', before => File['/tmp/a'] }
       file { '/tmp/a': ensure => 'present' }
       EOS
         end