+++ tasks/install_shell.sh
@@ -39,11 +39,14 @@ exists() {
 assert_unmodified_apt_config() {
   puppet_list=/etc/apt/sources.list.d/puppet.list
   puppet6_list=/etc/apt/sources.list.d/puppet6.list
+  puppet7_list=/etc/apt/sources.list.d/puppet7.list
 
   if [[ -f $puppet_list ]]; then
     list_file=puppet_list
   elif [[ -f $puppet6_list ]]; then
     list_file=puppet6_list
+  elif [[ -f $puppet7_list ]]; then
+    list_file=puppet7_list
   fi
 
   # If puppet.list exists, get its md5sum on disk and its md5sum from the puppet-release package
@@ -117,6 +120,12 @@ else
   else
     mac_source='http://downloads.puppet.com'
   fi
+fi
+
+if [ -n "$PT_retry" ]; then
+  retry=$PT_retry
+else
+  retry=5
 fi
 
 # Track to handle puppet5 to puppet6
@@ -301,10 +310,36 @@ capture_tmp_stderr() {
 
 trap "rm -f $tmp_stderr; rm -rf $tmp_dir; exit $1" 1 2 15
 
+# Run command and retry on failure
+# run_cmd CMD
+run_cmd() {
+  eval $1
+  rc=$?
+
+  if test $rc -ne 0; then
+    attempt_number=0
+    while test $attempt_number -lt $retry; do
+      info "Retrying... [$((attempt_number + 1))/$retry]"
+      eval $1
+      rc=$?
+
+      if test $rc -eq 0; then
+        break
+      fi
+
+      info "Return code: $rc"
+      sleep 1s
+      ((attempt_number=attempt_number+1))
+    done
+  fi
+
+  return $rc
+}
+
 # do_wget URL FILENAME
 do_wget() {
   info "Trying wget..."
-  wget -O "$2" "$1" 2>$tmp_stderr
+  run_cmd "wget -O '$2' '$1' 2>$tmp_stderr"
   rc=$?
 
   # check for 404
@@ -326,8 +361,9 @@ do_wget() {
 # do_curl URL FILENAME
 do_curl() {
   info "Trying curl..."
-  curl -1 -sL -D $tmp_stderr "$1" > "$2"
+  run_cmd "curl -1 -sL -D $tmp_stderr '$1' > '$2'"
   rc=$?
+
   # check for 404
   grep "404 Not Found" $tmp_stderr 2>&1 >/dev/null
   if test $? -eq 0; then
@@ -347,17 +383,31 @@ do_curl() {
 # do_fetch URL FILENAME
 do_fetch() {
   info "Trying fetch..."
-  fetch -o "$2" "$1" 2>$tmp_stderr
-  # check for bad return status
-  test $? -ne 0 && return 1
+  run_cmd "fetch -o '$2' '$1' 2>$tmp_stderr"
+  rc=$?
+
+  # check for 404
+  grep "404 Not Found" $tmp_stderr 2>&1 >/dev/null
+  if test $? -eq 0; then
+    critical "ERROR 404"
+    unable_to_retrieve_package
+  fi
+
+  # check for bad return status or empty output
+  if test $rc -ne 0 || test ! -s "$2"; then
+    capture_tmp_stderr "fetch"
+    return 1
+  fi
+
   return 0
 }
 
 # do_perl URL FILENAME
 do_perl() {
   info "Trying perl..."
-  perl -e 'use LWP::Simple; getprint($ARGV[0]);' "$1" > "$2" 2>$tmp_stderr
+  run_cmd "perl -e 'use LWP::Simple; getprint(\$ARGV[0]);' '$1' > '$2' 2>$tmp_stderr"
   rc=$?
+
   # check for 404
   grep "404 Not Found" $tmp_stderr 2>&1 >/dev/null
   if test $? -eq 0; then
@@ -428,9 +478,9 @@ install_file() {
 
       rpm -Uvh --oldpackage --replacepkgs "$2"
       if test "$version" = 'latest'; then
-        yum install -y puppet-agent && yum upgrade -y puppet-agent
+        run_cmd "yum install -y puppet-agent && yum upgrade -y puppet-agent"
       else
-        yum install -y "puppet-agent-${puppet_agent_version}"
+        run_cmd "yum install -y 'puppet-agent-${puppet_agent_version}'"
       fi
       ;;
     "noarch.rpm")
@@ -449,11 +499,11 @@ install_file() {
         fi
       fi
 
-      zypper install --no-confirm "$2"
+      run_cmd "zypper install --no-confirm '$2'"
       if test "$version" = "latest"; then
-        zypper install --no-confirm "puppet-agent"
+        run_cmd "zypper install --no-confirm 'puppet-agent'"
       else
-        zypper install --no-confirm --oldpackage --no-recommends --no-confirm "puppet-agent-${puppet_agent_version}"
+        run_cmd "zypper install --no-confirm --oldpackage --no-recommends --no-confirm 'puppet-agent-${puppet_agent_version}'"
       fi
       ;;
     "deb")
@@ -475,15 +525,15 @@ install_file() {
       assert_unmodified_apt_config
 
       dpkg -i --force-confmiss "$2"
-      apt-get update -y
+      run_cmd 'apt-get update -y'
 
       if test "$version" = 'latest'; then
-        apt-get install -y puppet-agent
+        run_cmd "apt-get install -y puppet-agent"
       else
         if test "x$deb_codename" != "x"; then
-          apt-get install -y "puppet-agent=${puppet_agent_version}-1${deb_codename}"
+          run_cmd "apt-get install -y 'puppet-agent=${puppet_agent_version}-1${deb_codename}'"
         else
-          apt-get install -y "puppet-agent=${puppet_agent_version}"
+          run_cmd "apt-get install -y 'puppet-agent=${puppet_agent_version}'"
         fi
       fi
       ;;
@@ -510,10 +560,12 @@ info "Downloading Puppet $version for ${platform}..."
 case $platform in
   "SLES")
     info "SLES platform! Lets get you an RPM..."
-    gpg_key="${tmp_dir}/RPM-GPG-KEY-puppet"
-    do_download "https://yum.puppet.com/RPM-GPG-KEY-puppet" "$gpg_key"
-    rpm --import "$gpg_key"
-    rm -f "$gpg_key"
+    for key in "puppet" "puppet-20250406"; do
+      gpg_key="${tmp_dir}/RPM-GPG-KEY-${key}"
+      do_download "https://yum.puppet.com/RPM-GPG-KEY-${key}" "$gpg_key"
+      rpm --import "$gpg_key"
+      rm -f "$gpg_key"
+    done
     filetype="noarch.rpm"
     filename="${collection}-release-sles-${platform_version}.noarch.rpm"
     download_url="${yum_source}/${filename}"